backtester: make data package more strict, add tests.

+++ backtester/data/data_types.go
@@ -45,7 +45,7 @@ type Holder interface {
 type Base struct {
 	latest     Event
 	stream     []Event
-	offset     int64
+	offset     int
 	isLiveData bool
 	m          sync.Mutex
 }
